work on UserController,InvoiceController,SizeController,modify some models,and work on CustomerRepository,InvoiceRepository,ProductRepository,SizeRepository, and CustomerService,UserService,InvoiceItemsService,InvoiceService,ProductService,SizeService

+++ src/main/kotlin/com/technolearn/rasoulonlineshop/repositories/invoices/InvoiceRepository.kt
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository
 @Repository
 interface InvoiceRepository : PagingAndSortingRepository<Invoice, Long>, CrudRepository<Invoice, Long> {
     @Query("from Invoice where user.id = :userId")
-    fun findAllByUserId(userId: Int, pageable: Pageable): List<Invoice>
+    fun findAllByUserId(userId: Long, pageable: Pageable): List<Invoice>
 
 }
